backend/api-gateway/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
flask-sock==0.7.0
PyJWT==2.8.0
redis==5.0.1
requests==2.31.0
